Refactor: Rearrange utils and added container.py
This is commit 3 of refactoring to move modules from utils to classes

commands.py had too many modules that were operations on a container
rather than operations on the command library dictionary or
dictionary containing information about the packages. Eventually,
those dictionaries should be migrated to classes and modules in
commands.py should be incorporated into those classes. Meanwhile,
this is a good middle ground.

- Moved container related modules into new file container.py
- Moved general modules into new file general.py
- Added some new modules to metadata.py
- Consolidated all the constants into constants.py
- Added 'WORKDIR' to list of docker directives

This commit will break the build
