+++ utils/commands.py
@@ -3,45 +3,19 @@ Copyright (c) 2017 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: BSD-2-Clause
 '''
 
-
-import grp
-import io
 import logging
 import os
-import pwd
-import re
 import subprocess
-import tarfile
-import time
 import yaml
 
-from contextlib import contextmanager
-
-import utils.constants as const
+from .container import docker_command
+from .container import execute
+from .general import parse_command
+from .constants import container
 
 '''
-Shell and docker command parser and invoking of commands
-within and outside a docker container
+Invoking commands in the command library
 '''
-# docker commands
-check_images = ['docker', 'images']
-pull = ['docker', 'pull']
-build = ['docker', 'build']
-run = ['docker', 'run', '-td']
-check_running = ['docker', 'ps', '-a']
-copy = ['docker', 'cp']
-execute = ['docker', 'exec']
-inspect = ['docker', 'inspect']
-stop = ['docker', 'stop']
-remove = ['docker', 'rm']
-delete = ['docker', 'rmi', '-f']
-save = ['docker', 'save']
-
-# docker container names
-# TODO: randomly generated image and container names
-image = 'tern-image'
-tag = str(int(time.time()))
-container = 'tern-container'
 
 # base image command library
 base_file = 'command_lib/base.yml'
@@ -74,77 +48,6 @@ def get_shell_commands(run_comm):
     for comm in comm_list:
         cleaned_list.append(comm.strip())
     return cleaned_list
-
-
-# from https://stackoverflow.com/questions/6194499/pushd-through-os-system
-@contextmanager
-def pushd(path):
-    curr_path = os.getcwd()
-    os.chdir(path)
-    yield
-    os.chdir(curr_path)
-
-
-def docker_command(command, *extra):
-    '''Invoke docker command. If the command fails nothing is returned
-    If it passes then the result is returned'''
-    full_cmd = []
-    sudo = True
-    try:
-        members = grp.getgrnam('docker').gr_mem
-        if pwd.getpwuid(os.getuid()).pw_name in members:
-            sudo = False
-    except KeyError:
-        pass
-    if sudo:
-        full_cmd.append('sudo')
-    full_cmd.extend(command)
-    for arg in extra:
-        full_cmd.append(arg)
-    # invoke
-    logger.debug("Running command: " + ' '.join(full_cmd))
-    pipes = subprocess.Popen(full_cmd, stdout=subprocess.PIPE,
-                             stderr=subprocess.PIPE)
-    result, error = pipes.communicate()
-    if error:
-        raise subprocess.CalledProcessError(1, cmd=full_cmd, output=error)
-    else:
-        return result
-
-
-def parse_command(command):
-    '''Typically a unix command is of the form:
-        command (subcommand) [options] [arguments]
-    Convert a given command into a dictionary of the form:
-        {'name': command,
-         'subcommand': subcommand,
-         'options': [list of options]
-         'arguments': [list of arguments]}'''
-    options = re.compile('^-')
-    options_list = []
-    args_list = []
-    command_dict = {}
-    command_words = command.split(' ')
-    # first word is the command name
-    command_dict.update({'name': command_words.pop(0), 'subcommand': ''})
-    # check if the first word is an option
-    first = command_words.pop(0)
-    if not options.match(first):
-        # this is a subcommand
-        command_dict.update({'subcommand': first})
-    else:
-        # append this to the list of options
-        options_list.append(first)
-    # find options and arguments in the rest of the list
-    while command_words:
-        if options.match(command_words[0]):
-            options_list.append(command_words.pop(0))
-        else:
-            args_list.append(command_words.pop(0))
-    # now we have options and arguments
-    command_dict.update({'options': options_list,
-                         'arguments': args_list})
-    return command_dict
 
 
 def check_sourcable(command, package_name):
@@ -199,8 +102,8 @@ def get_packages_per_run(docker_run_command):
     return pkg_dict
 
 
-def get_package_listing(docker_commands):
-    '''Given the docker commands in a dockerfile,  get a dictionary of
+def get_package_listing(docker_instructions):
+    '''Given the docker instructions in a dockerfile,  get a dictionary of
     packages that are in the command library of retrievable sources
     If it does not exist in the library then record them under
     unrecognized commands
@@ -212,9 +115,9 @@ def get_package_listing(docker_commands):
     '''
     pkg_dict = {'recognized': {}, 'unrecognized': []}
     shell_commands = []
-    for docker_command in docker_commands:
-        if docker_command[0] == 'RUN':
-            shell_commands.extend(get_shell_commands(docker_command[1]))
+    for instr in docker_instructions:
+        if instr[0] == 'RUN':
+            shell_commands.extend(get_shell_commands(instr[1]))
     for command in shell_commands:
         installed_dict = {'installed': [], 'removed': []}
         command_obj = parse_command(command)
@@ -253,68 +156,6 @@ def remove_uninstalled(pkg_dict):
                 installed_list.remove(remove)
         pkg_dict['recognized'].update({command: installed_list})
     return pkg_dict
-
-
-def check_container():
-    '''Check if a container exists'''
-    is_container = False
-    keyvalue = 'name=' + container
-    result = docker_command(check_running, '--filter', keyvalue)
-    result_lines = result.decode('utf-8').split('\n')
-    if len(result_lines) > 2:
-        is_container = True
-    return is_container
-
-
-def check_image(image_tag_string):
-    '''Check if image exists'''
-    is_image = False
-    result = docker_command(check_images, image_tag_string)
-    result_lines = result.decode('utf-8').split('\n')
-    if len(result_lines) > 2:
-        is_image = True
-    return is_image
-
-
-def build_container(dockerfile, image_tag_string):
-    '''Invoke docker command to build a docker image from the dockerfile
-    It is assumed that docker is installed and the docker daemon is running'''
-    curr_path = os.getcwd()
-    path = os.path.dirname(dockerfile)
-    if not check_image(image_tag_string):
-        with pushd(path):
-            try:
-                docker_command(build, '-t', image_tag_string, '-f',
-                               os.path.basename(dockerfile), '.')
-            except subprocess.CalledProcessError as error:
-                os.chdir(curr_path)
-                raise subprocess.CalledProcessError(
-                    error.returncode, cmd=error.cmd,
-                    output=error.output.decode('utf-8'))
-
-
-def start_container(image_tag_string):
-    '''Invoke docker command to start a container
-    If one already exists then stop it
-    Use this only in the beginning of running commands within a container
-    Assumptions: Docker is installed and the docker daemon is running
-    There is no other running container from the given image'''
-    if check_container():
-        remove_container()
-    docker_command(run, '--name', container, image_tag_string)
-
-
-def remove_container():
-    '''Remove a running container'''
-    if check_container():
-        docker_command(stop, container)
-        docker_command(remove, container)
-
-
-def remove_image(image_tag_string):
-    '''Remove an image'''
-    if check_image(image_tag_string):
-        docker_command(delete, image_tag_string)
 
 
 def get_base_info(image_tuple):
@@ -392,24 +233,3 @@ def get_pkg_attr_list(shell, attr_dict, package_name='', override=''):
                 else:
                     attr_list.append(result)
     return attr_list
-
-
-def get_image_id(image_tag_string):
-    '''Get the image ID by inspecting the image'''
-    result = docker_command(inspect, "-f'{{json .Id}}'", image_tag_string)
-    return result.split(':').pop()
-
-
-def extract_image_metadata(image_tag_string):
-    '''Run docker save and extract the files in a temporary directory'''
-    success = True
-    temp_path = os.path.abspath(const.temp_folder)
-    result = docker_command(save, image_tag_string)
-    if not result:
-        success = False
-    else:
-        with tarfile.open(fileobj=io.BytesIO(result)) as tar:
-            tar.extractall(temp_path)
-        if not os.path.exists(temp_path):
-            success = False
-    return success
