+++ utils/constants.py
@@ -9,3 +9,11 @@ Constants
 
 # temporary folder for extracting docker image
 temp_folder = 'temp'
+# built image name
+image = 'testimage'
+# running container name
+container = 'testcontainer'
+# logger name
+logger_name = 'docker-utils'
+# manifest file
+manifest_file = 'manifest.json'
