+++ utils/metadata.py
@@ -2,27 +2,24 @@
 Copyright (c) 2017 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: BSD-2-Clause
 '''
-
-
 import json
 import os
 import shutil
 
-from utils.commands import pushd
-import utils.constants as const
+from .general import pushd
+from .constants import temp_folder
+from .constants import manifest_file
+
 '''
-Docker metadata related modules
+Container metadata related modules
 NOTE: these modules work on a temp folder to which the output of docker save
 has already been extracted into
 '''
 
-# docker manifest file
-manifest_file = 'manifest.json'
-
 
 def clean_temp():
     '''Remove the temp directory'''
-    temp_path = os.path.abspath(const.temp_folder)
+    temp_path = os.path.abspath(temp_folder)
     if os.path.exists(temp_path):
         shutil.rmtree(temp_path)
 
@@ -30,7 +27,7 @@ def clean_temp():
 def get_image_manifest():
     '''Assuming that there is a temp folder with a manifest.json of
     an image inside, get a dict of the manifest.json file'''
-    temp_path = os.path.abspath(const.temp_folder)
+    temp_path = os.path.abspath(temp_folder)
     with pushd(temp_path):
         with open(manifest_file) as f:
             json_obj = json.loads(f.read())
@@ -39,12 +36,22 @@ def get_image_manifest():
 
 def get_image_layers(manifest):
     '''Given the manifest, return the layers'''
-    return manifest[0].get('Layers')
+    layers = []
+    for layer in manifest[0].get('Layers'):
+        layers.append(layer)
+    return layers
 
 
 def get_image_config_file(manifest):
     '''Given the manifest, return the config file'''
     return manifest[0].get('Config')
+
+
+def get_image_id(manifest):
+    '''Given the manifest, return the image id
+    This happens to be the config file's sha256sum'''
+    config_file = get_image_config_file(manifest)
+    return config_file.split('.')[0]
 
 
 def get_image_repotags(manifest):
@@ -58,13 +65,13 @@ def get_layer_sha(layer_path):
     return os.path.dirname(layer_path)
 
 
-def get_image_config():
+def get_image_config(manifest):
     '''Assuming there now exists a working directory where the image
     metadata exists, return the image config'''
-    config_file = get_image_config_file(get_image_manifest())
+    config_file = get_image_config_file(manifest)
     # assuming that the config file path is in the same root path as the
     # manifest file
-    temp_path = os.path.abspath(const.temp_folder)
+    temp_path = os.path.abspath(temp_folder)
     with pushd(temp_path):
         with open(config_file) as f:
             json_obj = json.loads(f.read())
@@ -80,7 +87,19 @@ def get_nonempty_history(config):
             if 'created_by' in item.keys():
                 history.append(item['created_by'])
             else:
-                history.append(item['comment'])
+                history.append('')
+    return history
+
+
+def get_history(config):
+    '''Given the image config, return the 'created_by' data. If it doesn't
+    exist return an empty string'''
+    history = []
+    for item in config['history']:
+        if 'created_by' in item.keys():
+            history.append(item['created_by'])
+        else:
+            history.append('')
     return history
 
 
@@ -90,3 +109,16 @@ def get_diff_ids(config):
     for item in config['rootfs']['diff_ids']:
         diff_ids.append(item.split(':').pop())
     return diff_ids
+
+
+def get_empty_diff_ids(config):
+    '''Given the image config, return a list of filesystem diff ids
+    if the config has an empty layer then return an empty string'''
+    diff_ids = get_diff_ids(config)
+    all_layers = []
+    for item in config['history']:
+        if 'empty_layer' not in item.keys():
+            all_layers.append(diff_ids.pop(0))
+        else:
+            all_layers.append('')
+    return all_layers
